backend/requirements-raw.txt
# prod
Django
django-environ
djangorestframework
django-allauth
jwt

# dev
requests
pytest
pytest-django
